Moved DistributorService and DistributorServiceImpl to new package and changed constructor parameter in MovieController
