+++ src/main/java/com/moviesdb/moviesdb/services/nonhuman/DistributorService.java
@@ -1,4 +1,4 @@
-package com.moviesdb.moviesdb.services.human;
+package com.moviesdb.moviesdb.services.nonhuman;
 
 import com.moviesdb.moviesdb.models.Distributor;
 import com.moviesdb.moviesdb.models.Movie;
